Creates 'Loop' class
